fix serialization tests

+++ experiments/data_loader.py
@@ -52,11 +52,13 @@ class NpyDatasetLoader(DatasetLoader):
             start_index, end_index = self.range
             train_dataset = np.load(self.train_dataset_path)[
                 start_index:end_index
-            ]
+            ].astype(np.float32, copy=False)
         else:
-            train_dataset = np.load(self.train_dataset_path)
-        queries = np.load(self.queries_path)
-        ground_truth = np.load(self.ground_truth_path)
+            train_dataset = np.load(self.train_dataset_path).astype(
+                np.float32, copy=False
+            )
+        queries = np.load(self.queries_path).astype(np.float32, copy=False)
+        ground_truth = np.load(self.ground_truth_path).astype(np.int32, copy=False)
         return train_dataset, queries, ground_truth
